requirements.txt
firecrawl-py
openai
pyyaml
git+https://github.com/openai/swarm.git
